MGeneralize components and move to common module
